requirements.txt
python-telegram-bot>=13.0,<15.0
schedule>=1.2.1
pickledb>=0.9